requirements.txt
mysql-connector==2.1.6
PyQt5==5.11.3
PyQt5-sip==4.19.13
